Move policy routes to permissions module and clean up actor resource
